move worker commons unit test from integration tests to worker file and do some improvements
