+++ node/core/pvf/common/src/worker.rs
@@ -201,6 +201,9 @@ pub mod thread {
 		thread,
 		time::Duration,
 	};
+	use std::sync::mpsc::channel;
+	use cpu_time::ProcessTime;
+	use crate::worker::cpu_time_monitor_loop;
 
 	/// Contains the outcome of waiting on threads, or `Pending` if none are ready.
 	#[derive(Clone, Copy)]
@@ -308,5 +311,71 @@ pub mod thread {
 		} else {
 			Some(*result.0)
 		}
+	}
+
+	#[test]
+	fn get_condvar_should_be_pending() {
+		let condvar = get_condvar();
+		let outcome = *condvar.0.lock().unwrap();
+		assert!(outcome.is_pending());
+	}
+
+	#[test]
+	fn wait_for_threads_with_timeout_return_none_on_time_out() {
+		let condvar = Arc::new((Mutex::new(WaitOutcome::Pending), Condvar::new()));
+		let outcome = wait_for_threads_with_timeout(&condvar, Duration::from_secs(1));
+		assert!(matches!(outcome, None));
+	}
+
+	#[test]
+	fn wait_for_threads_with_timeout_returns_outcome() {
+		let condvar = Arc::new((Mutex::new(WaitOutcome::Pending), Condvar::new()));
+		let condvar2 = Arc::clone(&condvar);
+		thread::spawn(move || {
+			cond_notify_all(Arc::clone(&condvar2), WaitOutcome::Finished);
+		});
+		let outcome = wait_for_threads_with_timeout(&condvar, Duration::from_secs(2));
+		assert!(matches!(outcome.unwrap(), WaitOutcome::Finished));
+	}
+
+	#[test]
+	fn spawn_worker_thread_should_notify_on_done() {
+		let condvar = Arc::new((Mutex::new(WaitOutcome::Pending), Condvar::new()));
+		let response =
+			spawn_worker_thread("thread", move || 2, Arc::clone(&condvar), WaitOutcome::TimedOut);
+		let (lock, _) = &*condvar;
+		let r = response.unwrap().join().unwrap();
+		assert!(matches!(r, 2));
+		assert!(matches!(*lock.lock().unwrap(), WaitOutcome::TimedOut));
+	}
+
+	#[test]
+	fn spawn_worker_thread_should_not_notify() {
+		let condvar = Arc::new((Mutex::new(WaitOutcome::Finished), Condvar::new()));
+		let response =
+			spawn_worker_thread("thread", move || 2, Arc::clone(&condvar), WaitOutcome::TimedOut);
+		let (lock, _) = &*condvar;
+		let r = response.unwrap().join().unwrap();
+		assert!(matches!(r, 2));
+		assert!(matches!(*lock.lock().unwrap(), WaitOutcome::Finished));
+	}
+
+	#[test]
+	fn cpu_time_monitor_loop_should_return_time_elapsed() {
+		let cpu_time_start = ProcessTime::now();
+		let timeout = Duration::from_secs(0);
+		let (_tx, rx) = channel();
+		let result = cpu_time_monitor_loop(cpu_time_start, timeout, rx);
+		assert_ne!(result, None);
+	}
+
+	#[test]
+	fn cpu_time_monitor_loop_should_return_none() {
+		let cpu_time_start = ProcessTime::now();
+		let timeout = Duration::from_secs(10);
+		let (tx, rx) = channel();
+		tx.send(()).unwrap();
+		let result = cpu_time_monitor_loop(cpu_time_start, timeout, rx);
+		assert_eq!(result, None);
 	}
 }
